requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn

python-dotenv

# LiveKit 1.x
livekit-agents[deepgram,openai,elevenlabs,silero,turn-detector]~=1.1
# Si besoin : livekit-plugins-assemblyai~=1.1

# LlamaIndex & embeddings
llama-index==0.10.68
llama-index-embeddings-huggingface
sentence-transformers

# (optionnel) Ollama local
llama-index-llms-ollama

# Google Calendar API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dateparser

